backend/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
bitarray==2.8.0
certifi==2023.7.22
charset-normalizer==3.2.0
cytoolz==0.12.2
deprecation==2.1.0
eth-abi==4.1.0
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.2.0
frozenlist==1.4.0
gotrue==1.0.3
h11==0.14.0
hexbytes==0.3.1
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
lru-dict==1.2.0
multidict==6.0.4
onesignal-python-api==2.0.2
packaging==23.1
parsimonious==0.9.0
pluggy==1.2.0
postgrest==0.10.8
protobuf==4.23.4
pycryptodome==3.18.0
pydantic==2.1.1
pydantic_core==2.4.0
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pyunormalize==15.0.0
realtime==1.0.0
referencing==0.30.2
regex==2023.6.3
requests==2.31.0
rlp==3.0.0
rpds-py==0.9.2
six==1.16.0
sniffio==1.3.0
storage3==0.5.4
StrEnum==0.4.15
supabase==1.0.4
supafunc==0.2.3
toolz==0.12.0
typing_extensions==4.7.1
urllib3==2.0.4
web3==6.8.0
websockets==10.4
yarl==1.9.2
